+++ tmc-langs-ant/pom.xml
@@ -18,13 +18,6 @@
     <artifactId>tmc-langs-ant</artifactId>
     <packaging>jar</packaging>
 
-    <repositories>
-        <repository>
-            <id>testmycode.net</id>
-            <url>http://maven.testmycode.net/nexus/content/repositories/releases/</url>
-        </repository>
-    </repositories>
-
     <dependencies>
         <dependency>
             <groupId>fi.helsinki.cs.tmc</groupId>
