+++ tmc-langs-ant/src/main/java/fi/helsinki/cs/tmc/langs/ant/AntPlugin.java
@@ -39,53 +39,6 @@ public class AntPlugin extends LanguagePluginAbstract {
         return "apache-ant";
     }
 
-    /**
-     *
-     * @param basePath The file path to search in.
-     * @return A list of directories that contain a build file in this language.
-     */
-    @Override
-    public ImmutableList<Path> findExercises(Path basePath) {
-        File searchPath = new File(basePath.toString());
-        ImmutableList.Builder<Path> listBuilder = new ImmutableList.Builder<>();
-        if (searchPath.exists() && searchPath.isDirectory()) {
-            return search(searchPath, listBuilder);
-        } else {
-            return listBuilder.build();
-        }
-    }
-
-    /**
-     * Search a directory and its subdirectories for build files. If a directory
-     * contains a build file, the directory is added to the list.
-     *
-     * @param file The current file path to search in
-     * @param listBuilder
-     * @return a list of all directories that contain build files for this
-     * language.
-     */
-    private ImmutableList<Path> search(File file, ImmutableList.Builder<Path> listBuilder) {
-        Stack<File> stack = new Stack();
-        // Push the initial directory onto the stack.
-        stack.push(file);
-        // Walk the directories that get added onto the stack.
-        while (!stack.isEmpty()) {
-            File current = stack.pop();
-            if (current.isDirectory()) {
-                // See if current directory contains a build file.
-                if (isExerciseTypeCorrect(current.toPath())) {
-                    listBuilder.add(current.toPath());
-                }
-                for (File temp : current.listFiles()) {
-                    if (temp.isDirectory()) {
-                        stack.push(temp);
-                    }
-                }
-            }
-        }
-        return listBuilder.build();
-    }
-
     @Override
     public ExerciseDesc scanExercise(Path path, String exerciseName) {
         if (!isExerciseTypeCorrect(path)) {
@@ -151,62 +104,12 @@ public class AntPlugin extends LanguagePluginAbstract {
 
     @Override
     public RunResult runTests(Path path) {
-        if (!isExerciseTypeCorrect(path)) {
-            throw new RuntimeException("Project has no build.xml");
-        } else {
-            build(path);
-            String name = path.getFileName().toString().split(".")[0];
-            File file = path.toFile();
-            File dir = file.getParentFile();
-            URLClassLoader classLoader = null;
-            try {
-                URL url = new URL("file://" + dir.getAbsolutePath());
-                classLoader = new URLClassLoader(new URL[]{url});
-            } catch (MalformedURLException ex) {
-                Logger.getLogger(AntPlugin.class.getName()).log(Level.SEVERE, null, ex);
-            }
-
-            Class<?> clazz = null;
-            try {
-                clazz = classLoader.loadClass(name);
-            } catch (ClassNotFoundException ex) {
-                Logger.getLogger(AntPlugin.class.getName()).log(Level.SEVERE, null, ex);
-            }
-
-            Result testResults = JUnitCore.runClasses(clazz);
-
-            return parseAndConvertTestResults(testResults);
-
-        }
-    }
-
-    private void build(Path path) {
-        Project project = new Project();
-        File buildFile = new File(path.toString() + File.separatorChar + "build.xml");
-        project.setUserProperty("ant.file", buildFile.getAbsolutePath());
-        project.init();
-
-        ProjectHelper helper = ProjectHelper.getProjectHelper();
-        project.addReference("ant.projectHelper", helper);
-        helper.parse(project, buildFile);
-
-        project.executeTarget(project.getDefaultTarget());
-    }
-
-    /**
-     * Check if the exercise's project type corresponds with the language plugin
-     * type.
-     *
-     * @param path The path to the exercise directory.
-     * @return True if given path is valid directory for this language plugin
-     */
-    @Override
-    public boolean isExerciseTypeCorrect(Path path) {
-        return new File(path.toString() + File.separatorChar + "build.xml").exists();
-    }
-
-    private RunResult parseAndConvertTestResults(Result testResults) {
         return null;
+    }
+
+    @Override
+    protected boolean isExerciseTypeCorrect(Path path) {
+        return new File(path.toString() + File.separatorChar + "build.xml").exists();
     }
     
     @Override
