+++ tmc-langs-framework/pom.xml
@@ -19,17 +19,12 @@
     <name>tmc-langs-framework</name>
     
     <url>http://testmycode.net</url>
-    <repositories>
-        <repository>
-            <id>testmycode.net</id>
-            <url>http://maven.testmycode.net/nexus/content/repositories/releases/</url>
-        </repository>
-    </repositories>
+
     <dependencies>
         <dependency>
             <groupId>fi.helsinki.cs.tmc</groupId>
             <artifactId>tmc-checkstyle-runner</artifactId>
-            <version>1.0.3-SNAPSHOT</version>    
+            <version>1.0.2</version>    
         </dependency>
         <dependency>
             <groupId>com.google.guava</groupId>
