+++ tmc-langs-framework/src/main/java/fi/helsinki/cs/tmc/langs/LanguagePluginAbstract.java
@@ -6,7 +6,10 @@
 
 package fi.helsinki.cs.tmc.langs;
 
+import com.google.common.collect.ImmutableList;
+import java.io.File;
 import java.nio.file.Path;
+import java.util.Stack;
 
 public abstract class LanguagePluginAbstract implements LanguagePlugin {
     
@@ -30,5 +33,61 @@ public abstract class LanguagePluginAbstract implements LanguagePlugin {
     @Override
     public void prepareSolution(Path path) {
         exerciseBuilder.prepareSolution(path);
+    }
+    
+    /**
+     * Check if the exercise's project type corresponds with the language plugin
+     * type.
+     *
+     * @param path The path to the exercise directory.
+     * @return True if given path is valid directory for this language plugin
+     */
+    protected abstract boolean isExerciseTypeCorrect(Path path);
+    
+    /**
+     *
+     * @param basePath The file path to search in.
+     * @return A list of directories that contain a build file in this language.
+     */
+    @Override
+    public ImmutableList<Path> findExercises(Path basePath) {
+        File searchPath = new File(basePath.toString());
+        ImmutableList.Builder<Path> listBuilder = new ImmutableList.Builder<>();
+        if (searchPath.exists() && searchPath.isDirectory()) {
+            return search(searchPath, listBuilder);
+        } else {
+            return listBuilder.build();
+        }
+    }
+
+    /**
+     * Search a directory and its subdirectories for build files. If a directory
+     * contains a build file, the directory is added to the list.
+     *
+     * @param file The current file path to search in
+     * @param listBuilder
+     * @return a list of all directories that contain build files for this
+     * language.
+     */
+    private ImmutableList<Path> search(File file, ImmutableList.Builder<Path> listBuilder) {
+        Stack<File> stack = new Stack();
+        // Push the initial directory onto the stack.
+        stack.push(file);
+        // Walk the directories that get added onto the stack.
+        while (!stack.isEmpty()) {
+            File current = stack.pop();
+            if (current.isDirectory()) {
+                // See if current directory contains a build file.
+                if (isExerciseTypeCorrect(current.toPath())) {
+                    listBuilder.add(current.toPath());
+                }
+                for (File temp : current.listFiles()) {
+                    if (temp.isDirectory()) {
+                        stack.push(temp);
+                    }
+                }
+            }
+        }
+        return listBuilder.build();
     }
 }
